requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.5
python-multipart==0.0.20
sqlalchemy==2.0.36
aiosqlite==0.21.0
apscheduler==3.11.0
httpx==0.28.1
python-dotenv==1.0.1
